refactor: move positioning
